med_llm/requirements.txt
flask~=3.0.0
functions~=0.7.0
openai~=1.2.3
packaging~=23.2
requests~=2.31.0
waitress
python-dotenv==1.0.1
